Update sam4s_hal.c
On the SAM4S board, the silkscreen labels indicate PA14-LED3, PA15-LED2, and PA16-LED1. However, the led_ok and led_error functions are not implemented for this microcontroller, unlike other microcontrollers such as the STM32

As a result, when running notebooks that use these functions on the SAM4S, the LEDs remain off, as the calls to led_ok and led_error are ignored.

This pull request implements led_ok(status) and led_error(status) in sam4s_hal.c, ensuring that these LEDs function correctly when executing notebooks that rely on them. This provides a consistent visual indication across different microcontrollers.